refactor: handleNewBooking #3
