+++ packages/features/bookings/lib/handleSeats/reschedule/owner/moveSeatedBookingToNewTimeSlot.ts
@@ -6,8 +6,9 @@ import { sendRescheduledEmails } from "@calcom/emails";
 import prisma from "@calcom/prisma";
 import type { AdditionalInformation, AppsStatus } from "@calcom/types/Calendar";
 
-import { addVideoCallDataToEvent, findBookingQuery } from "../../../handleNewBooking";
+import { addVideoCallDataToEvent } from "../../../handleNewBooking";
 import type { Booking, createLoggerWithEventDetails } from "../../../handleNewBooking";
+import { findBookingQuery } from "../../../handleNewBooking/findBookingQuery";
 import { handleAppsStatus } from "../../../handleNewBooking/handleAppsStatus";
 import type { SeatedBooking, RescheduleSeatedBookingObject } from "../../types";
