+++ packages/lib/defaultEvents.ts
@@ -167,3 +167,5 @@ export const getUsernameList = (users: string | string[] | undefined): string[] 
 };
 
 export default defaultEvents;
+
+export type AwaitedGetDefaultEvent = Awaited<ReturnType<typeof getDefaultEvent>>;
